Refactor download code
