+++ app/service/download.py
@@ -1,6 +1,5 @@
 """Functions to support browsing the RDS document structure"""
 
-import os
 import zipfile
 from io import BytesIO, StringIO
 from logging import getLogger
@@ -11,27 +10,8 @@ from fastapi import Depends
 
 from app.clients.db.session import get_db
 from app.repository.download import get_whole_database_dump
-from app.repository.helpers import get_query_template
 
 _LOGGER = getLogger(__name__)
-
-
-def create_query(
-    template_query, ingest_cycle_start: str, allowed_corpora_ids: list[str]
-) -> str:
-    """Create download whole database query, replacing variables.
-
-    :param str ingest_cycle_start: The current ingest cycle date.
-    :param list[str] allowed_corpora_ids: The corpora from which we
-        should allow the data to be dumped.
-    :return str: The SQL query to perform on the database session.
-    """
-    corpora_ids = "'" + "','".join(allowed_corpora_ids) + "'"
-    return template_query.replace(  # type: ignore
-        "{ingest_cycle_start}", ingest_cycle_start
-    ).replace(
-        "{allowed_corpora_ids}", corpora_ids
-    )  # type: ignore
 
 
 def replace_slug_with_qualified_url(
@@ -63,8 +43,10 @@ def convert_dump_to_csv(df: pd.DataFrame):
     return csv_buffer
 
 
-def generate_data_dump_as_csv(query, db=Depends(get_db)):
-    df = get_whole_database_dump(query, db)
+def generate_data_dump_as_csv(
+    ingest_cycle_start: str, allowed_corpora_ids: list[str], db=Depends(get_db)
+):
+    df = get_whole_database_dump(ingest_cycle_start, allowed_corpora_ids, db)
     csv = convert_dump_to_csv(df)
     csv.seek(0)
     return csv
@@ -92,11 +74,8 @@ def create_data_download_zip_archive(
     ingest_cycle_start: str, allowed_corpora_ids: list[str], db=Depends(get_db)
 ):
     readme_buffer = generate_data_dump_readme(ingest_cycle_start)
-    query_template = get_query_template(
-        os.path.join("app", "repository", "sql", "download.sql")
-    )
-    query = create_query(query_template, ingest_cycle_start, allowed_corpora_ids)
-    csv_buffer = generate_data_dump_as_csv(query, db)
+
+    csv_buffer = generate_data_dump_as_csv(ingest_cycle_start, allowed_corpora_ids, db)
 
     zip_buffer = BytesIO()
     with zipfile.ZipFile(zip_buffer, "a", zipfile.ZIP_DEFLATED, False) as zip_file:
